requirements.txt
absl-py==2.0.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appnope==0.1.3
asgiref==3.9.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
better-profanity==0.7.0
billiard==4.2.1
black==23.11.0
blinker==1.9.0
blis==0.7.9
boto3==1.26.109
botocore==1.29.111
CacheControl==0.13.1
cachetools==5.3.1
catalogue==2.0.8
celery==5.5.3
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.6.2
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.0.0
comm==0.2.1
confection==0.0.4
contourpy==1.2.0
cryptography==41.0.1
cycler==0.12.1
cymem==2.0.7
debugpy==1.8.0
decorator==5.1.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@a0e22dbfa791e6235e4f196d5ce25e754d02be31
distlib==0.3.7
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1
django-redis==6.0.0
django-tailwind==4.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
filelock==3.12.2
firebase-admin==6.1.0
fonttools==4.46.0
fsspec==2023.6.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
googletrans==4.0.0rc1
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.15.1
hydra-core==1.3.2
hyperframe==5.2.0
idna==2.10
iopath==0.1.9
ipdb==0.13.11
ipykernel==6.29.2
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
kombu==5.5.4
langcodes==3.3.0
langdetect==1.0.9
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
murmurhash==1.0.9
mypy-extensions==1.0.0
narwhals==1.24.2
nest-asyncio==1.6.0
networkx==3.1
nltk==3.7
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
outcome==1.2.0
packaging==23.1
pandas==2.2.3
parso==0.8.3
pathspec==0.11.2
pathy==0.10.1
pexpect==4.8.0
phonenumbers==8.13.8
pickleshare==0.7.5
pika==1.3.2
Pillow==9.4.0
pipenv==2023.7.23
platformdirs==3.10.0
portalocker==2.8.2
preshed==3.0.8
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.3
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyarrow==19.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.0.2
pydantic_core==2.1.2
pydeck==0.9.1
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0
pyzmq==25.1.2
redis==6.4.0
referencing==0.36.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.6.1
selenium==4.11.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.5.3
srsly==2.4.6
stack-data==0.6.2
starlette==0.27.0
streamlit==1.41.1
sympy==1.12
tabulate==0.9.0
Telethon==1.28.1
tenacity==9.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
thinc==8.1.10
tokenizers==0.13.3
toml==0.10.2
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
trio==0.22.2
trio-websocket==0.10.3
typer==0.7.0
typing_extensions==4.12.2
tzdata==2025.2
undetected-chromedriver==3.5.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
vine==5.1.0
virtualenv==20.24.2
virtualenv-clone==0.5.7
wasabi==1.1.2
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==3.0.1
wget==3.2
wsproto==1.2.0
yacs==0.1.8
